requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-jet-reboot==1.1.1
djangorestframework==3.12.4
google-api-python-client==1.4.1
httplib2==0.19.1
oauth2client==4.1.3
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1